# backend/validation/openapi_be13_generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "verifAI / RefCheck AI Backend",
    "description": "Backend foundation for the verifAI / RefCheck AI Generative AI course project.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/references/{reference_id}": {
      "get": {
        "tags": [
          "references"
        ],
        "summary": "Reference Details",
        "operationId": "reference_details_api_v1_references__reference_id__get",
        "parameters": [
          {
            "name": "reference_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reference Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/references/{reference_id}/verify-doi": {
      "post": {
        "tags": [
          "references"
        ],
        "summary": "Verify Reference Doi",
        "operationId": "verify_reference_doi_api_v1_references__reference_id__verify_doi_post",
        "parameters": [
          {
            "name": "reference_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reference Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/references/{reference_id}/metadata": {
      "get": {
        "tags": [
          "references"
        ],
        "summary": "Reference Metadata",
        "operationId": "reference_metadata_api_v1_references__reference_id__metadata_get",
        "parameters": [
          {
            "name": "reference_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reference Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/{document_id}/summary": {
      "get": {
        "tags": [
          "reports-feedback-uat"
        ],
        "summary": "Document Summary",
        "operationId": "document_summary_api_v1_documents__document_id__summary_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/reports": {
      "post": {
        "tags": [
          "reports-feedback-uat"
        ],
        "summary": "Generate Document Report",
        "operationId": "generate_document_report_api_v1_documents__document_id__reports_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/GenerateReportRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/report": {
      "get": {
        "tags": [
          "reports-feedback-uat"
        ],
        "summary": "Get Latest Document Report",
        "operationId": "get_latest_document_report_api_v1_documents__document_id__report_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "tags": [
          "reports-feedback-uat"
        ],
        "summary": "Get Report",
        "operationId": "get_report_api_v1_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}/download": {
      "get": {
        "tags": [
          "reports-feedback-uat"
        ],
        "summary": "Download Report",
        "operationId": "download_report_api_v1_reports__report_id__download_get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "HTML",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verification-results/{result_id}/feedback": {
      "post": {
        "tags": [
          "reports-feedback-uat"
        ],
        "summary": "Submit Verification Feedback",
        "operationId": "submit_verification_feedback_api_v1_verification_results__result_id__feedback_post",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/claim-reference-links/{link_id}/feedback": {
      "post": {
        "tags": [
          "reports-feedback-uat"
        ],
        "summary": "Submit Claim Reference Mapping Feedback",
        "operationId": "submit_claim_reference_mapping_feedback_api_v1_claim_reference_links__link_id__feedback_post",
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Link Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MappingFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uat/surveys": {
      "post": {
        "tags": [
          "reports-feedback-uat"
        ],
        "summary": "Submit Uat Survey",
        "operationId": "submit_uat_survey_api_v1_uat_surveys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UatSurveyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GenerateReportRequest": {
        "properties": {
          "format": {
            "type": "string",
            "title": "Format",
            "default": "HTML"
          },
          "include_evidence_chunks": {
            "type": "boolean",
            "title": "Include Evidence Chunks",
            "default": true
          },
          "include_human_review_items": {
            "type": "boolean",
            "title": "Include Human Review Items",
            "default": true
          },
          "include_limitations": {
            "type": "boolean",
            "title": "Include Limitations",
            "default": true
          }
        },
        "type": "object",
        "title": "GenerateReportRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MappingFeedbackRequest": {
        "properties": {
          "feedback_type": {
            "type": "string",
            "title": "Feedback Type",
            "default": "OTHER"
          },
          "suggested_reference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Reference Id"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "user_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Role"
          }
        },
        "type": "object",
        "title": "MappingFeedbackRequest"
      },
...
      "UatSurveyRequest": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "participant_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participant Role"
          },
          "ease_of_use_rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Ease Of Use Rating"
          },
          "result_clarity_rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Result Clarity Rating"
          },
          "trust_rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Trust Rating"
          },
          "usefulness_rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Usefulness Rating"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "ease_of_use_rating",
          "result_clarity_rating",
          "trust_rating",
          "usefulness_rating"
        ],
        "title": "UatSurveyRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationFeedbackRequest": {
        "properties": {
          "user_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Label"
          },
          "user_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Comment"
          },
          "user_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Role"
          }
        },
        "type": "object",
        "title": "VerificationFeedbackRequest"
      }
    }
  }
}